+++ src/CommandManager.php
@@ -14,27 +14,80 @@ class CommandManager {
 
   protected Configuration $configuration;
 
+  /**
+   * Constructs a Command manager object.
+   */
   public function __construct() {
     $this->configuration = new Configuration();
   }
 
-  public function getCommands(): array {
-    $config_dir = $this->configuration->getConfigDir();
-    $config = $this->configuration->loadConfigFile();
-    $command_configs = $this->configuration->loadCommandConfigFiles();
+  /**
+   * Returns global commands.
+   *
+   * @return array
+   *   The array of commands objects.
+   */
+  public function getGlobalCommands(): array {
+    $global_config_dir = $this->configuration->getGlobalConfigDir();
+    // Unfortunately realpath doesn't handle '~' in path.
+    if (str_starts_with($global_config_dir, '~')) {
+      $global_config_dir = $_SERVER['HOME'] . substr($global_config_dir, 1);
+    }
 
+    return $this->getCommandsFromFolder($global_config_dir);
+  }
+
+  /**
+   * Returns local commands.
+   *
+   * @return array
+   *   The array of commands objects.
+   */
+  public function getLocalCommands(): array {
+    $config_dir = $this->configuration->getConfigDir();
+    if (!$config_dir) {
+      return [];
+    }
+    return $this->getCommandsFromFolder($config_dir);
+  }
+
+  /**
+   * Returns combined global and local commands.
+   *
+   * @return array
+   *   The array of commands objects.
+   */
+  public function getCommands(): array {
+    $global_commands = $this->getGlobalCommands();
+    $local_commands = $this->getLocalCommands();
+    $commands = array_merge($global_commands, $local_commands);
+    return array_values($commands);
+  }
+
+  /**
+   * Returns commands for a given folder.
+   *
+   * @param string $config_dir
+   *   The path to folder for scanning for commands.
+   *
+   * @return array
+   *   The array of commands objects.
+   */
+  public function getCommandsFromFolder(string $config_dir): array {
+    $commands = [];
+    $config = $this->configuration->loadConfigFile();
     // Load env files.
     if (isset($config['env_file'])) {
       $dotenv = new Dotenv();
       foreach ($config['env_file'] as $env_file) {
-        $env_file = realpath($config_dir . '/' . $env_file);
+        $env_file = realpath($this->configuration->getConfigDir() . '/' . $env_file);
         if ($env_file) {
           $dotenv->load($env_file);
         }
       }
     }
 
-    $commands = [];
+    $command_configs = $this->configuration->loadCommandConfigFiles($config_dir);
     foreach ($command_configs as $command_config) {
       $command = new Command();
       $command->setName($command_config['name']);
@@ -103,7 +156,7 @@ class CommandManager {
         return Command::SUCCESS;
       });
 
-      $commands[] = $command;
+      $commands[$command->getName()] = $command;
     }
 
     return $commands;